server/requirements.txt
asgiref==3.5.2
async-timeout==4.0.2
attrs==22.1.0
backports.zoneinfo==0.2.1
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.1.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==37.0.4
defusedxml==0.7.1
Deprecated==1.2.13
Django==4.1
django-templated-mail==1.1.1
djangorestframework==3.13.1
djangorestframework-simplejwt==4.8.0
djoser==2.1.0
drf-spectacular==0.23.1
drf-yasg==1.21.3
idna==3.3
importlib-resources==5.9.0
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
jsonschema==4.11.0
MarkupSafe==2.1.1
oauthlib==3.2.0
packaging==21.3
pkgutil_resolve_name==1.3.10
pycparser==2.21
PyJWT==2.1.0
pyparsing==3.0.9
pyrsistent==0.18.1
python3-openid==3.2.0
pytz==2022.2.1
PyYAML==6.0
redis==4.3.4
requests==2.28.1
requests-oauthlib==1.3.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
six==1.16.0
social-auth-app-django==4.0.0
social-auth-core==4.3.0
sqlparse==0.4.2
tzdata==2022.2
uritemplate==4.1.1
urllib3==1.26.11
wrapt==1.14.1
zipp==3.8.1
